+++ packages/circuits/scripts/build_witness_generator.sh
@@ -6,5 +6,10 @@ CIRCUIT_NAME="subtreeupdate"
 CIRCUIT_ARTIFACTS_DIR="$ROOT_DIR/circuit-artifacts/"
 
 pushd "$CIRCUIT_ARTIFACTS_DIR"/"$CIRCUIT_NAME"/"$CIRCUIT_NAME"_cpp
+# clean
+rm *.o
+rm "$CIRCUIT_NAME"
+
+# build
 make
 popd